in autoAug, an unnecessary error message is now avoided. checkParamArchive.pl is used by the webserver application.
git-svn-id: file:///home/gumz3/Desktop/svnAugustusRepo/svnArchiveFile/svn/augustus/trunks@84 946bced8-b911-4d30-a262-e0a769a8c5ad

+++ scripts/autoAug.pl
@@ -1004,10 +1004,11 @@ sub collect{
     my $configDir="$AUGUSTUS_CONFIG_PATH/species/$species";
     print '3 cd ../config'."\n" if ($verbose>=3);
     chdir "../config";
-    $cmdString = "cp $configDir/* . ; rm *.orig*;";
-    print "3 $cmdString\n" if ($verbose>=3);
-    system("$cmdString")==0 or die ("failed to execute: $cmdString\n");
-    
+    if(-e "*.orig*"){
+       $cmdString = "cp $configDir/* . ; rm *.orig*;";
+       print "3 $cmdString\n" if ($verbose>=3);
+       system("$cmdString")==0 or die ("failed to execute: $cmdString\n");
+    }
     # collect files with gb format
 
     print "3 cd ../genes\n" if ($verbose>=3);
